moved image related actions to their own file
